[RPC tests] Capture subprocess output and only print it on failure
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45439

RPC tests are quite verbose (and, later in this stack, will become even more so) which makes it hard to navigate the logs to find the relevant part for a failing test. This PR redirects the stdout and stderr of child processes (i.e., which is where we run RPCs from) to some files, and only copies the content of these files to the "real" stdout if the test fails (otherwise the files are ignored). This means that successful tests won't print anything.

There were two options to do this: use a single file for all subprocesses' stdout and stderr, so that it would all get merged into a single "stream" (which is basically the effect we had before), or use separate files so that the logs of the various processes can be kept separate. The former option has the benefit of preserving the relative order of messages of different processes, while the latter option avoids ambiguity and conflicts. (In fact, many log lines are already tagged with a timestamp and a PID, so we can reconstruct some of that information after the fact). In the end, I went for the latter option, but let me know if the former is preferable.
ghstack-source-id: 113032407

Differential Revision: [D23963476](https://our.internmc.facebook.com/intern/diff/D23963476/)