[RPC tests] Have GDB print the stack traces of all threads upon deadlock
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45442

Some of our tests are flaky but when they fail they print very little information which makes it hard to diagnose the issue. One notable instance of this occurs with deadlocks, where the processes stop, don't print any logs, and are eventually killed by the test runner, leaving few hints as to where they were and why they weren't progressing. One useful piece of information to understand deadlocks comes from TensorPipe's verbose logging trace, which we enabled in an earlier PR in this stack. Another useful tool is having the stack traces of all the threads at the time the deadlock occurred. One easy way of obtaining it is to use GDB, which is what this PR here does. (If GDB isn't installed it fails graciously).
ghstack-source-id: 113032406

Differential Revision: [D23963479](https://our.internmc.facebook.com/intern/diff/D23963479/)